+++ plugins/ocp/ocp-utils.c
@@ -13,7 +13,7 @@ const unsigned char ocp_uuid[NVME_UUID_LEN] = {
 	0xc1, 0x94, 0xd5, 0x5b, 0xe0, 0x94, 0x47, 0x94, 0xa2, 0x1d,
 	0x29, 0x99, 0x8f, 0x56, 0xbe, 0x6f };
 
-int ocp_find_uuid_index(struct nvme_id_uuid_list *uuid_list, int *index)
+int ocp_find_uuid_index(struct nvme_id_uuid_list *uuid_list, __u8 *index)
 {
 	int i = nvme_uuid_find(uuid_list, ocp_uuid);
 
@@ -26,7 +26,7 @@ int ocp_find_uuid_index(struct nvme_id_uuid_list *uuid_list, int *index)
 	return 0;
 }
 
-int ocp_get_uuid_index(struct nvme_dev *dev, int *index)
+int ocp_get_uuid_index(struct nvme_dev *dev, __u8 *index)
 {
 	struct nvme_id_uuid_list uuid_list;
 	int err = nvme_identify_uuid(dev_fd(dev), &uuid_list);
